Testing on real mouse data
Wrote code to remove repeated positions in the from_file data.
Changed the maps to make the obstacle smaller, to avoid invalid moves.
Changed q learning runner a bit so that it accepts file data with 1 or 2 sets of coordinates (the code for 1 set of coordinates was broken).
Added in real mouse training data for condition 1.